airflow_3.1.0_requirements.txt
# Airflow 3.1.0 전체 의존성 목록
# 폐쇄망 환경에서 사용하기 위한 패키지 리스트

# ============================================================
# 1. Airflow 코어
# ============================================================
apache-airflow==3.1.0

# ============================================================
# 2. 핵심 의존성
# ============================================================
alembic>=1.13.1
argcomplete>=1.10
asgiref>=3.5.2
attrs>=22.1.0
blinker>=1.6.2
colorlog>=6.7.0
configupdater>=3.1.1
connexion[flask]>=3.0.0
cron-descriptor>=1.2.24
croniter>=2.0.2
cryptography>=41.0.0
deprecated>=1.2.13
dill>=0.3.2
flask>=2.2.5,<3.1
flask-appbuilder>=4.3.0
flask-caching>=2.1.0
flask-login>=0.6.2
flask-session>=0.5.0
flask-wtf>=1.1.0
graphviz>=0.12
gunicorn>=20.1.0
httpx>=0.25.0
importlib-metadata>=6.5.0
importlib-resources>=6.1.0
itsdangerous>=2.0
jinja2>=3.1.2
jsonschema>=4.18.0
lazy-object-proxy
linkify-it-py>=2.0.0
lockfile>=0.12.2
markdown>=3.0
markdown-it-py>=2.1.0
markupsafe>=2.1.2
marshmallow>=3.19.0
marshmallow-oneofschema>=3.0.1
mdit-py-plugins>=0.3.0
methodtools>=0.4.7
opentelemetry-api>=1.15.0
opentelemetry-exporter-otlp>=1.15.0
packaging>=23.0
pathspec>=0.9.0
pendulum>=3.0.0
pluggy>=1.0
psutil>=5.8.0
pygments>=2.0.1
pyjwt>=2.4.0
python-daemon>=3.0.0
python-dateutil>=2.3
python-nvd3>=0.15.0
python-slugify>=5.0
pytz>=2021.3
pyyaml>=5.4
requests>=2.27.0
rfc3339-validator>=0.1.4
rich>=13.0.0
rich-argparse>=1.0.0
setproctitle>=1.1.10
sqlalchemy>=1.4.36,<2.0
sqlalchemy-jsonfield>=1.0
tabulate>=0.7.5
tenacity>=8.0.0
termcolor>=1.1.0
typing-extensions>=4.0.0
unicodecsv>=0.14.1
werkzeug>=2.0,<4.0
wtforms>=2.3.3,<4.0

# ============================================================
# 3. 데이터베이스 드라이버
# ============================================================
# PostgreSQL (권장)
psycopg2-binary>=2.9.0

# MySQL (선택)
# mysqlclient>=1.3.6

# SQLite (Python 내장, 별도 설치 불필요)

# ============================================================
# 4. Airflow Providers (선택적)
# ============================================================
# HTTP
apache-airflow-providers-http

# PostgreSQL
apache-airflow-providers-postgres

# Amazon (S3, EMR 등)
apache-airflow-providers-amazon

# Google Cloud
apache-airflow-providers-google

# Slack
apache-airflow-providers-slack

# Email
apache-airflow-providers-email

# FTP
apache-airflow-providers-ftp

# SSH
apache-airflow-providers-ssh

# ============================================================
# 5. 추가 유틸리티 (선택적)
# ============================================================
# Celery Executor
celery>=5.3.0

# Redis (Celery 백엔드)
redis>=4.5.0

# Flower (Celery 모니터링)
flower>=2.0.0

# ============================================================
# 6. C to Java 프로젝트 의존성
# ============================================================
# HTTP 클라이언트 (watsonx, GitHub API)
requests>=2.31.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 암호화 (이미 위에 cryptography>=41.0.0 포함됨)

# Git 작업 (subprocess 사용, 별도 패키지 불필요)
# Git CLI가 시스템에 설치되어 있어야 함

# IBM watsonx.ai SDK (선택적, requests로 직접 API 호출 가능)
# ibm-watson>=7.0.0
# ibm-cloud-sdk-core>=3.16.0

# ============================================================
# 7. 개발 도구 (선택적)
# ============================================================
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
